improve integrations page, add delete option

+++ app/integrations/mutations/deleteIntegration.ts
@@ -1,3 +1,4 @@
+import Guard, { restrictEditToOwnDocument } from "app/guard/ability"
 import { resolver } from "blitz"
 import db from "db"
 import { z } from "zod"
@@ -9,9 +10,10 @@ const DeleteIntegration = z.object({
 export default resolver.pipe(
   resolver.zod(DeleteIntegration),
   resolver.authorize(),
-  async ({ id }) => {
-    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-    const integration = await db.integration.deleteMany({ where: { id } })
+  restrictEditToOwnDocument(),
+  Guard.authorizePipe("delete:own", "Integration"),
+  async ({ id, requiredUserId }) => {
+    const integration = await db.integration.deleteMany({ where: { id, userId: requiredUserId } })
 
     return integration
   }
